// src/app/utils/protobuf/CatMessage.d.ts
export namespace Cat {

    /** Properties of a Base. */
    interface IBase {

        /** Base Id */
        Id?: (string|null);

        /** Base Sender */
        Sender?: (string|null);

        /** Base Receiver */
        Receiver?: (string|null);

        /** Base SendOn */
        SendOn?: (string|null);
    }
// ...
    interface IChat {

        /** Chat Info */
        Info?: (Cat.IBase|null);

        /** Chat Body */
        Body?: (string|null);
    }
// ...
    enum MessageType {
        LOGIN = 0,
        PING = 1,
        CHAT = 2,
        IMG = 3,
        FILE = 4,
        ADD_FRIEND = 1001
    }
// ...
}